+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
  
 ## 2.0.0
 - improved dependencies
+- added multiactiontag-functionality
 
 ### new features
 - none
@@ -11,13 +12,15 @@
 ### improvements
 - bumped up and improved dependencies
 - added CommonDocActionTagService
+- added multiactiontag-functionality with header-component and add behavior to searchpage, inlinesearchpaage, list, list-item (moved from mymediacollection)
  
 ### bug fixes
 - fixed tests
  
 ### breaking changes
-- new constructor for CommonDocActionsComponent
-
+- changed constructor for CommonDocActionsComponent
+- changed constructor for CommonDocSearchPageComponent
+- changed constructor for CommonDocInlineSearchpageComponent
 
 ## 1.0.0
 - initial version based on mytourbook-1.5.0